==> notebooks/data/05_persist_pixels_and_binary_masks.ipynb <==
# Test creating a binary mask

import geopandas as gpd
import leafmap
from shapely.ops import unary_union
from shapely.geometry import Point, mapping, box, shape
import shapely
import json
import os
import rasterio
import matplotlib.pyplot as plt
import numpy as np

import sys
sys.path.append("..")

os.getcwd()
root = os.path.dirname(os.getcwd())
print(root)

from src.data.get_satellite_images import ReadSTAC

post_dataset = root + "/data/processed/mining_tiles_with_masks_and_bounding_boxes.gpkg"
print(gpd.list_layers(post_dataset))
tiles = gpd.read_file(post_dataset, layer="tiles")
masks = gpd.read_file(post_dataset, layer="preferred_polygons")
masks

# select first row
row = tiles.iloc[0]
row

bounds = row.geometry.bounds
bounds


# get lat and lon
lat = row.geometry.centroid.y
lon = row.geometry.centroid.x
lat, lon

# Download the image 
api_url="https://planetarycomputer.microsoft.com/api/stac/v1"
bands = ['B04', 'B03', 'B02']

stac_reader = ReadSTAC(api_url=api_url, collection = "sentinel-2-l2a")

item = stac_reader.get_item_by_name(row.s2_tile_id, bbox=bounds)

# read as stack
stack = stac_reader.get_stack(
    items=item, 
    bands=bands,
    crop_to_bounds=False, 
    squeeze_time_dim=True,
    custom_point_and_buffer=[lon, lat, 10240],
    chunk_size=512,
    )

# persist to disk
# stac_reader.save_stack_as_geotiff(stack, root + f"/data/processed/images/{row.tile_id}_{row.s2_tile_id}_img.tif")

# show the stack 
stac_reader.display_stack_as_image(stack)

# read the mask 
tile_id = row.tile_id
poly = masks[masks.tile_id == tile_id].geometry.values

# convert to same crs as stack
poly = poly.to_crs(stack.crs)

mask_raster = rasterio.features.rasterize(poly, out_shape=(2048,2048), transform=stack.rio.transform())

np.unique(mask_raster)

# show the image
plt.imshow(mask_raster)

# Testing (Delete)

import rasterio
from rasterio.features import rasterize
from shapely.geometry import Polygon

# Define your Shapely polygon (example polygon in EPSG:32620)
polygon = Polygon([(500000, 500000), (500000, 510000), (510000, 510000), (510000, 500000)])

# Define the spatial resolution and bounds of your output raster
transform = rasterio.transform.from_origin(west=500000, north=510000, xsize=10, ysize=10)
width, height = 2048, 2048  # dimensions of the raster

# Define the output raster's CRS
crs = "EPSG:32620"

# Rasterize the polygon
binary_mask = rasterize(
    [(polygon, 1)],
    out_shape=(height, width),
    transform=transform,
    fill=0,  # background value
    all_touched=True,
    dtype='uint8'
)

# binary_mask now contains the binary mask of the rasterized polygon
np.unique(binary_mask)

plt.imshow(binary_mask)
